refactor(build): archive deployment profile
This changes the Maven configuration to remove the
`deploy.s2i.repository` and `deploy.ui.distribution` and instead rely on
the `deploy.archives` property. Which in turn activates the
`deploy-archives` profiles that include the assembly and deployment
Maven plugin configuration.

To build and deploy the build configuration needs to specify the
`-Ddeploy.archives=true` via command line, that way it's a system
property and it can activate the `deploy-archives` profile.