feat(operator): postgresql database upgrade
We changed the version of the PostgreSQL database from 9.5 to 9.6. Any
time a a non patch version of the PostgreSQL database is increased
`pg_upgrade` must be run on the data directory (with old and new
binaries) to upgrade the data files.

Luckily, the Software Collections PostgreSQL image already contains
 script that checks for `POSTGRESQL_UPGRADE` environment variable and if
it's set to either `copy` or `hardlink` it will perform run the
`pg_upgrade` on startup. Caveat being that the `POSTGRESQL_UPGRADE` can
be only set if the upgrade is required; meaning that we can't set it and
leave it we have to manage it: set when needed and removed when not.

This adds a `DatabaseNeedsUpgrade` flag in the configuration and
depending on that flag sets the `POSTGRESQL_UPGRADE=copy` environment
variable. The `DatabaseNeedsUpgrade` flag is set if the container image
of the current `postgresql` container defined in the `DeploymentConfig`
is different from the image configured via
`Syndesis.Components.Database.Image` configuration.

On installation the state of the `syndesis-db` deployment is watched and
when it reaches the point where specified number of replicas equals the
ready number of replicas the `DeploymentConfig` is patched to remove the
`POSTGRESQL_UPGRADE=copy` environment variable.

This will cause a temporary outage, as changing the `DeploymentConfig`
results in the creation of new replication controller and termination of
the currently running `syndesis-db` pod.